moved class search handler to correct location
