+++ tests/Unit/Extension/SourceCodeFilesystem/Rpc/ClassSearchHandlerTest.php
@@ -3,7 +3,7 @@
 namespace Phpactor\Tests\Unit\Extension\SourceCodeFilesystem\Rpc;
 
 use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilestem\Application\ClassSearch;
-use Phpactor\Extension\Rpc\Handler\ClassSearchHandler;
+use Phpactor\Extension\SourceCodeFilesystem\Rpc\ClassSearchHandler;
 use Phpactor\Extension\Rpc\Handler;
 use Phpactor\Extension\Rpc\Response\EchoResponse;
 use Phpactor\Extension\Rpc\Response\ReturnResponse;
